fix(loan-ai): 5 bugs found in deep audit
1. CRITICAL: Move /simulate route BEFORE /{loan_id} to prevent
   FastAPI treating 'simulate' as a loan_id path parameter
2. Move mid-file imports to top (PEP8 compliance)
3. Fix person_income type: int → float (match ORM Numeric(18,2))
4. Move 'import pandas' to module level (avoid per-request overhead)
5. Strip num__/cat__ prefixes from metadata fallback feature names

+++ backend/app/schemas/loan.py
@@ -120,7 +120,7 @@ class LoanListItem(BaseModel):
 class LoanSimulationRequest(BaseModel):
     """Input parameters for Loan AI Simulation (predict PD Score)."""
     person_age: int = Field(..., ge=18, le=100, description="Age of the applicant")
-    person_income: int = Field(..., ge=0, description="Annual income")
+    person_income: float = Field(..., ge=0, description="Annual income")
     person_home_ownership: str = Field(..., description="RENT, MORTGAGE, OWN, OTHER")
     person_emp_length: int = Field(..., ge=0, description="Employment length in years")
     loan_intent: str = Field(..., description="PERSONAL, EDUCATION, MEDICAL, VENTURE, HOMEIMPROVEMENT, DEBTCONSOLIDATION")
